download/digital-human-livestream/requirements.txt
# Digital Human YouTube Livestream - Dependencies

# Web scraping and news
requests>=2.28.0
beautifulsoup4>=4.12.0
feedparser>=6.0.0

# Audio processing
pydub>=0.25.0
mutagen>=1.47.0

# Video processing
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0

# FFmpeg wrapper
ffmpeg-python>=0.2.0

# API and async
aiohttp>=3.8.0
asyncio>=3.4.3

# Logging and utilities
loguru>=0.7.0
python-dotenv>=1.0.0

# Schedule
schedule>=1.2.0
